requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
openai
apscheduler
pytz
